% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notie.R, R/use.R
\name{notie}
\alias{notie}
\alias{notie_close}
\alias{use_notie}
\alias{use_push}
\title{Notification with notie.js}
\usage{
notie(text, time = getOption("notie.time", default = 3),
  type = c("info", "neutral", "success", "error", "warning"),
  position = c("bottom", "top"),
  session = shiny::getDefaultReactiveDomain())

notie_close(session = shiny::getDefaultReactiveDomain())

use_notie()

use_push()
}
\arguments{
\item{text}{Text to display.}

\item{time}{Delay for closing notification in seconds (s).}

\item{type}{Type of notification: \code{"alert"}, \code{"success"}, \code{"error"}, \code{"warning"}, \code{"info"}.}

\item{position}{Where to display notification: \code{"bottom"} or \code{"top"}.}

\item{session}{Shiny session.}
}
\description{
Initialize with \code{use_notie} in UI before using \code{notie} server-side.
 Close all notifications with \code{notie_close}.
}
\examples{
if (interactive()) {



}
}
