% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notie.R
\name{notie_confirm}
\alias{notie_confirm}
\title{Confirmation pop-up with notie.js}
\usage{
notie_confirm(inputId, text, label_submit = "Ok",
  label_cancel = "Cancel", position = c("top", "bottom"),
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{text}{Text to display.}

\item{label_submit}{Submit button label.}

\item{label_cancel}{Cancel button label.}

\item{position}{Where to display notification: \code{"bottom"} or \code{"top"}.}

\item{session}{Shiny session.}
}
\description{
Initialize with \code{use_notie} in UI before using \code{notie_confirm} server-side.
 Retrieve value server-side with \code{input$<inputId>}.
}
\examples{
if (interactive()) {

  library(shiny)

  ui <- fluidPage(
    use_notie(),
    tags$h2("notie.js confirmation pop-up"),
    actionButton("show", "Ask for confirmation"),
    verbatimTextOutput("result")
  )

  server <- function(input, output, session) {

    observeEvent(input$show, {
      notie_confirm(
        inputId = "confirm",
        text = "Do you want to confirm?",
        label_submit = "Sure!",
        label_cancel = "Nope!"
      )
    })

    output$result <- renderPrint({
      input$confirm
    })
  }

  shinyApp(ui, server)

}
}
