% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noty.R, R/use.R
\name{noty}
\alias{noty}
\alias{noty_close}
\alias{use_noty}
\title{Notification with noty.js}
\usage{
noty(text, timeout = getOption("noty.timeout", default = 1000),
  type = c("info", "alert", "success", "error", "warning"),
  layout = c("topRight", "top", "topLeft", "topCenter", "center",
  "centerLeft", "centerRight", "bottom", "bottomLeft", "bottomCenter",
  "bottomRight"), theme = getOption("noty.theme", default = "sunset"),
  modal = FALSE, killer = FALSE, animation_open = NULL,
  animation_close = NULL, session = shiny::getDefaultReactiveDomain())

noty_close(session = shiny::getDefaultReactiveDomain())

use_noty()
}
\arguments{
\item{text}{Text to display.}

\item{timeout}{Delay for closing notification in milliseconds (ms).
Set \code{FALSE} for sticky notifications.}

\item{type}{Type of notification: \code{"alert"}, \code{"success"}, \code{"error"}, \code{"warning"}, \code{"info"}.}

\item{layout}{Position amongst : \code{"top"}, \code{"topLeft"}, \code{"topCenter"},
\code{"topRight"}, \code{"center"}, \code{"centerLeft"}, \code{"centerRight"},
\code{"bottom"}, \code{"bottomLeft"}, \code{"bottomCenter"}, "\code{"bottomRight"}.}

\item{theme}{Theme to use between : \code{"mint"}, \code{"sunset"}, \code{"relax"}, \code{"nest"}, \code{"metroui"},
\code{"semanticui"}, \code{"light"}, \code{"bootstrap-v3"}, \code{"bootstrap-v4"}.}

\item{modal}{Add an overlay to the page to emphasis notification.}

\item{killer}{Close all others notification before openning.}

\item{animation_open, animation_close}{Animation effect, use \code{\link{animations}}.}

\item{session}{Shiny session.}
}
\description{
Initialize with \code{use_noty} in UI before using \code{noty} server-side.
 Close all notifications with \code{noty_close}.
}
\examples{
if (interactive()) {



}
}
