% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Notification with push.js}
\usage{
push(title, text = NULL, timeout = 4000,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{title}{Title of the notification.}

\item{text}{Optionnal text of the notification}

\item{timeout}{Time in milliseconds until the notification closes automatically.}

\item{session}{Shiny session.}
}
\description{
Initialize with \code{use_push} in UI before using \code{push} server-side.
 Close all notifications with \code{push_close}.
}
\note{
push.js sends desktop notifications, user need to allow permission to show notification.
}
\examples{
if (interactive()) {

  library(shiny)

  ui <- fluidPage(
    tags$h2("Push notification"),
    use_push(),
    actionButton("show", "Show notification"),
    textInput("title", "Text to display", "Hello world :)")
  )

  server <- function(input, output, session) {

    observeEvent(input$show, {
      push(title = input$title)
    })

  }

  shinyApp(ui, server)

}
}
